ai/requirements.txt
langgraph
langchain
groq
gitpython
pydantic
python-dotenv